# apps/backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Our Haven API
  version: 0.0.0
  description: Our Haven v1 API — Parent + Provider mobile, Provider web portal,
    admin dashboard. Source of truth per ADR-0004 (§§ 1–3, 8) and ADR-0010
    (Supabase + Fly.io + Vercel platform). US-region only.
  license:
    name: Proprietary
components:
  securitySchemes:
    supabaseAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token (HS256). Verified locally with the
        project JWT secret on every request.
...
paths:
...
  /v1/uploads/signed-url:
    post:
      summary: Create a short-lived signed PUT URL for client-side Supabase Storage
        upload
      tags:
        - uploads
      description: Returns a short-lived (5 min default) Supabase Storage signed
        upload URL + token. The client uses `uploadToSignedUrl` (or a fetch PUT
        with the token) to send the file directly to Supabase Storage.
        Content-type and per-kind max size are validated on the request;
        bucket-level size caps in Supabase Storage backstop the limit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kind:
                  type: string
                  enum:
                    - id-doc
                    - license-doc
                    - insurance-doc
                    - state-childcare-registration
                    - avatar
                contentType:
                  type: string
                  enum:
                    - image/jpeg
                    - image/png
                    - application/pdf
                contentLengthBytes:
                  type: integer
                  exclusiveMinimum: 0
                  maximum: 20971520
              required:
                - kind
                - contentType
                - contentLengthBytes
      security:
        - supabaseAccessToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadUrl:
                    type: string
                    format: uri
                  uploadToken:
                    type: string
                  objectPath:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - uploadUrl
                  - uploadToken
                  - objectPath
                  - expiresAt
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  field:
                    type: string
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  field:
                    type: string
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: https://api.ourhaven.example
    description: Production (Fly.io iad)
  - url: http://localhost:8080
    description: Local development
...